formatting, decode, and format options components, and move things in animtab around a bit
